dotnet soap

+++ WS_EUREKABANK_SOAPDOTNET_GR04/02.CLICON/EB_DOTNET_SOAP_CliCon/ec.edu.monster.modelo/CliCon_Movimiento.cs
@@ -1,8 +1,4 @@
 ﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
 
 namespace EB_DOTNET_SOAP_CliCon.ec.edu.monster.modelo
 {
@@ -13,6 +9,6 @@ namespace EB_DOTNET_SOAP_CliCon.ec.edu.monster.modelo
         public DateTime Fecha { get; set; }
         public string Tipo { get; set; }
         public string Accion { get; set; }
-        public double Importe { get; set; }
+        public decimal Importe { get; set; } // <- antes double
     }
 }
